move the logic for muting users from the /users endpoint to the /mod one, since that's an action reserved to moderators and administrators
